refactor: consolidate shared OAuth client and rate-limit constants
Extract Store#discogs_oauth_client to eliminate duplicate OAuth client
construction in OrderPoller and CsvExport. Extract Discogs::RateLimit
module for shared backoff constants and logic, consumed by both
Marketplace and DiscogsRateLimitMiddleware. Skip delete_all query
in SoldListingRemover when no listings matched.
